feat(sovereignty): Rolling Biometric Evolution + supervisor fold + Tranche 2 (515 lines)
E2E Sovereignty transition (operator authorization 2026-07-19):

1. REAL enrollment wired (mandate 1 — zero hardcoded paths):
   biometric_evolution.py resolves the operator's ACTUAL profile
   ('Derek J. Russell', 192-dim speechbrain x-vector, 272 samples)
   through speaker_profile_store's OWN env-driven resolver.
   Mandate correction stated openly: the enrolled tensor is an
   X-VECTOR, not MFCCs — evolution blends in the profile's NATIVE
   embedding space (blending MFCCs into an x-vector store would
   corrupt it). Pipeline: normalize_acoustics (numpy pre-emphasis +
   log-spectral mean subtraction — the classic static-channel/reverb
   strip) → score → conf > JARVIS_SENTRY_EVOLUTION_DELTA (0.85) →
   slow-EMA blend (JARVIS_BIOMETRIC_EVOLUTION_ALPHA 0.02) → persisted
   to the SAME row, total_samples bumped. STRICT tensor guards: shape
   mismatch / non-finite blend ABORT — enrollment can never corrupt.
   Marginal samples NEVER teach. Wired into BiometricGateAdapter
   (normalization pre-filter before ANY scoring; evolution on
   clear-pass only).
2. Monolith fold: mount_passive_sentry now owns the main-runloop pump
   as an asyncio task on the supervisor's own loop — the ambient loop
   needs NO scratchpad host and survives native launchd ignition.
3. Tranche 2: automated veto pass over ALL 60 remaining Tier-A
   candidates → 22 verified clean, 38 vetoed (external consumers) →
   515 more lines migrated to quarantine/supervisor_slice2.py.
   TWO BUGS CAUGHT BY PROOF-BEFORE-SHIP: (a) the net map hardcoded
   slice1 — tranche-2 revivals would AttributeError (remapped +
   script generalized via QMOD); (b) extracted zones lost the
   monolith's import context (NameError: NamedTuple) — slice modules
   now carry a stdlib import header; slice2 imports CLEAN and both
   tranches revive through CHAINED PEP-562 nets (live-proven).
   Supervisor total shed: 1,314 lines across both tranches.

6 new tests incl. mandate-4 verbatim (high-confidence chunk →
baseline tensor updated in memory, exact EMA law, shape uncorrupted,
sample counted; marginal never teaches; mismatch aborts) + the REAL
enrollment resolving read-only. 103-test sweep green.


[integrity-verified: e08b2b0bd9a9]
